requirements.txt
Django==1.9.7
django-markdownx==1.6
django-material==0.8.0
Markdown==2.6.6
Pillow==3.2.0
